+++ collection_tests/list_test.c
@@ -7,7 +7,7 @@
 #include <string.h>
 
 #define dynlist_test(arr, list) \
-    for (size_t i = 0; i < list_length(list); i++) { \
+    for (int i = 0; i < list_length(list); i++) { \
         obs_test(get_data(list->data[i].data, arr), ==, arr[i]) \
     } \
 
@@ -18,7 +18,7 @@ int main(void) {
         OBS_TEST("Create list and test properties", {
             List list = list_new("1");
             obs_test_strcmp(list->parent.name, "1");
-            obs_test(list_length(list), ==, (size_t)0);
+            obs_test(list_length(list), ==, (int)0);
             obs_test_true(list_is_empty(list));
             list_free(list);
         })
@@ -78,10 +78,10 @@ int main(void) {
     OBS_TEST_GROUP("List_clear", {
         OBS_TEST("Clearing empty list", {
             List list = list_new("1");
-            obs_test(list_length(list), ==, (size_t)0);
+            obs_test(list_length(list), ==, (int)0);
             obs_test_true(list_is_empty(list));
             list_clear(list, false);
-            obs_test(list_length(list), ==, (size_t)0);
+            obs_test(list_length(list), ==, (int)0);
             obs_test_true(list_is_empty(list));
             list_free(list);
         })
@@ -92,7 +92,7 @@ int main(void) {
             map_items(list, 5, elements, list, list_push_back);
             dynlist_test(elements, list);
             list_clear(list, false);
-            obs_test(list_length(list), ==, (size_t)0);
+            obs_test(list_length(list), ==, (int)0);
             obs_test_true(list_is_empty(list));
             list_free(list);
         })
@@ -106,14 +106,14 @@ int main(void) {
     OBS_TEST_GROUP("List_insert_after/before", {
         OBS_TEST("Inserting after on empty list", {
             List list = list_new("1");
-            obs_test(list_length(list), ==, (size_t)0);
+            obs_test(list_length(list), ==, (int)0);
             obs_test_true(list_is_empty(list));
             list_free(list);
         })
 
         OBS_TEST("Inserting before on empty list", {
             List list = list_new("2");
-            obs_test(list_length(list), ==, (size_t)0);
+            obs_test(list_length(list), ==, (int)0);
             obs_test_true(list_is_empty(list));
             list_free(list);
         })
@@ -179,14 +179,14 @@ int main(void) {
     OBS_TEST_GROUP("list_length", {
         OBS_TEST("Empty list", {
             List list = list_new("1");
-            obs_test(list_length(list), ==, (size_t)0);
+            obs_test(list_length(list), ==, (int)0);
             list_free(list);
         })
 
         OBS_TEST("Single element list", {
             List list = list_new("2");
             list_push_back(list, NEW_NODE(list, 2));
-            obs_test(list_length(list), ==, (size_t)1);
+            obs_test(list_length(list), ==, (int)1);
             list_free(list);
         })
 
@@ -195,7 +195,7 @@ int main(void) {
             long long *items = ((long long[]){1, 2, 3, 4, 5, 6, 7, 8, 9, 10});
             map_items(list, 10, items, list, list_push_back);
             dynlist_test(items, list);
-            obs_test(list_length(list), ==, (size_t)10);
+            obs_test(list_length(list), ==, (int)10);
             list_free(list);
         })
     })
@@ -204,7 +204,7 @@ int main(void) {
         OBS_TEST("Append to empty list", {
             List list = list_new("1");
             list_push_back(list, NEW_NODE(list, 10));
-            obs_test(list_length(list), ==, (size_t)1);
+            obs_test(list_length(list), ==, (int)1);
             obs_test(list->data[0].data.int_data, ==, (long long)10);
             list_free(list);
         })
@@ -215,7 +215,7 @@ int main(void) {
             map_items(list, 7, elements, list, list_push_back);
             dynlist_test(elements, list);
             list_push_back(list, NEW_NODE(list, 10));
-            obs_test(list_length(list), ==, (size_t)8);
+            obs_test(list_length(list), ==, (int)8);
             obs_test(list->data[list->cur_len - 1].data.int_data, ==, (long long)10);
             list_free(list);
         })
