+++ src/general_collection_helper.h
@@ -6,13 +6,13 @@
 
 #define MAX_SIZE_FLT                    (15)
 
-size_t sizeof_uint(unsigned long long int n);
+int sizeof_uint(unsigned long long int n);
 
-size_t sizeof_int(long long int n);
+int sizeof_int(long long int n);
 
-void write_str_center_incr(wchar_t **buf, size_t *offset, size_t len,
-                           wchar_t *str, size_t str_len);
+void write_str_center_incr(wchar_t **buf, int *offset, int len,
+                           wchar_t *str, int str_len);
 
-size_t sizeof_data(Data data, TypeTag tag);
+int sizeof_data(Data data, TypeTag tag);
 
-#endif
+#endif /* LLV_GENERAL_COLLECTION_HELPER */
